pyproject.toml
[project]
name = "Gemini_Test_Path_Finder"
version = "0.1.0"
description = "An example project for the use of the Gemini LLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
